- move armor rendering
