+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## v1.1.27 (1.20.1)
 - fixed enchantments not modifying base attack damage during heavy attack
 - fixed shaped-mining not respecting enchantments
+- fixed slot positioning in UI
+- moved to fabrics armor rendering api for compat (this fixes cosmetic armor compat)
