+++ common/src/main/resources/miapi-common.mixins.json
@@ -43,7 +43,6 @@
         "ThrowablePotionItemAccessor"
     ],
     "client": [
-        "client.ArmorFeatureRendererMixin",
         "client.BuiltinModelItemRendererMixin",
         "client.ClientPlayerEntityMixin",
         "client.ElytraEntityModelAccessor",
