+++ fabric/src/main/java/smartin/miapi/fabric/MiapiFabric.java
@@ -3,6 +3,7 @@ package smartin.miapi.fabric;
 import com.jamieswhiteshirt.reachentityattributes.ReachEntityAttributes;
 import dev.architectury.platform.Platform;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.client.rendering.v1.ArmorRenderer;
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.fabric.api.resource.ResourceReloadListenerKeys;
@@ -49,6 +50,9 @@ public class MiapiFabric implements ModInitializer {
         }
 
         MiapiReloadListener listener = new MiapiReloadListener();
+        RegistryInventory.modularItems.addCallback((modularItem) -> {
+            ArmorRenderer.register(new ModularArmorRenderer(), modularItem);
+        });
 
         ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new IdentifiableResourceReloadListener() {
             @Override
