+++ gradle.properties
@@ -33,7 +33,7 @@ armory_version=1.0-pre-release-3-1.20.1
 
 #Fabric dependencies
 fabric_loader_version=0.15.6
-fabric_api_version=0.86.0+1.20.1
+fabric_api_version=0.92.2+1.20.1
 fabric_modmenu_version=7.2.1
 fabric_reach_range_version=2.4.0
 fabric_apoli_version=v2.11.2
